Completed ToDo's
Note: This code is very buggy and required a phone with a lot of memory to run - it will not run on Nexus 5 (Android 6.0).  On some phones the image must be taken in landscape mode.  Changing orientation after taking image can cause crash.

Added method getClassification() which returns the probability left eye is open closed, right eye is open/closed and if the face is smiling or not.